+++ eatngo-infra/src/main/kotlin/com/eatngo/common/BaseJpaEntity.kt
@@ -1,19 +1,29 @@
-package com.eatngo.order.rdb.entity.common
+package com.eatngo.common
 
 import jakarta.persistence.EntityListeners
 import jakarta.persistence.MappedSuperclass
 import org.springframework.data.annotation.CreatedBy
 import org.springframework.data.annotation.CreatedDate
+import org.springframework.data.annotation.LastModifiedBy
+import org.springframework.data.annotation.LastModifiedDate
 import org.springframework.data.jpa.domain.support.AuditingEntityListener
 import java.time.LocalDateTime
 
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener::class)
-abstract class CreatableJpaEntity(
-
+abstract class BaseJpaEntity(
     @CreatedBy
     var createdBy: String? = null,
 
     @CreatedDate
-    var createdAt: LocalDateTime? = null
+    var createdAt: LocalDateTime = LocalDateTime.now(),
+
+    @LastModifiedBy
+    var updatedBy: String? = null,
+
+    @LastModifiedDate
+    var updatedAt: LocalDateTime = LocalDateTime.now(),
+
+
+    var deletedAt: LocalDateTime? = null,
 )
