+++ eatngo-infra/src/main/kotlin/com/eatngo/order/rdb/repository/OrderRdbRepository.kt
@@ -3,5 +3,5 @@ package com.eatngo.order.rdb.repository
 import com.eatngo.order.rdb.entity.OrderJpaEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface OrderRdbRepository: JpaRepository<OrderJpaEntity, Long> {
+interface OrderRdbRepository : JpaRepository<OrderJpaEntity, Long> {
 }
