+++ eatngo-infra/src/main/kotlin/com/eatngo/store_owner/persistence/StoreOwnerPersistenceImpl.kt
@@ -1,11 +1,11 @@
-package com.eatngo.order.persistence
+package com.eatngo.store_owner.persistence
 
 import com.eatngo.store_owner.domain.StoreOwner
 import com.eatngo.store_owner.infra.StoreOwnerPersistence
 import org.springframework.stereotype.Component
 
 @Component
-class StoreOwnerPersistenceImpl: StoreOwnerPersistence {
+class StoreOwnerPersistenceImpl : StoreOwnerPersistence {
     override fun save(storeOwner: StoreOwner): StoreOwner {
         TODO("Not yet implemented")
     }
