+++ eatngo-infra/src/main/kotlin/com/eatngo/user_account/persistence/UserAccountPersistence.kt
@@ -1,21 +1,24 @@
-package com.eatngo.order.persistence
+package com.eatngo.user_account.persistence
 
 import com.eatngo.user_account.domain.UserAccount
 import com.eatngo.user_account.infra.UserAccountPersistence
 import org.springframework.stereotype.Component
 
 @Component
-class UserAccountPersistenceImpl : UserAccountPersistence {
+class UserAccountPersistenceImpl(
+) : UserAccountPersistence {
+
     override fun save(account: UserAccount): UserAccount {
-        TODO("Not yet implemented")
+        // TODO : Implement save logic
+        return account
     }
 
     override fun findById(id: Long): UserAccount? {
-        TODO("Not yet implemented")
+        // TODO : Implement findById logic
+        return null
     }
 
     override fun deleteById(id: Long) {
-        TODO("Not yet implemented")
+        // TODO : Implement deleteById logic
     }
-
 }
